web/backend/requirements.txt
langchain
langgraph
google-genai
python-dotenv
fastapi
uvicorn